Move non-gConsent specific files to common module
The common module only holds internal APIs shared between the gConsent module and the ODRL module.
