+++ src/gConsent/manage/getAccessGrantAll.ts
@@ -30,7 +30,7 @@ import type { AccessBaseOptions } from "../type/AccessBaseOptions";
 import type { RecursivePartial } from "../../type/RecursivePartial";
 import type { IssueAccessRequestParameters } from "../type/IssueAccessRequestParameters";
 import { accessToResourceAccessModeArray } from "../util/accessToResourceAccessModeArray";
-import { getSessionFetch } from "../util/getSessionFetch";
+import { getSessionFetch } from "../../common/util/getSessionFetch";
 import type { BaseGrantBody } from "../type/AccessVerifiableCredential";
 
 /**
